\name{forestChange-package}
\alias{forestChange-package}
\alias{forestChange}
\docType{package}
\title{Computing Essential Biodiversity Variables from Global Forest
Change Data}

\description{Metrics and statistics of Essential Biodiversity Variables are computed by processing Global Forest Change data (Hansen et al., 2013) <doi:10.1126/science.1244693> , Canopy Cover data (Sexton et al., 2013) <doi:10.1080/17538947.2013.786146>, and polygon geometries (e.g., GADM, see <https://gadm.org/>).}

\details{

The DESCRIPTION file:
\packageDESCRIPTION{forestChange}
\packageIndices{forestChange}
Maintainer: \packageMaintainer{forestChange}
}
\author{Wilson Lara <wilarhen@temple.edu>, Victor Gutierrez-Velez <victorhugo@temple.edu>}



\keyword{ package }


