\name{forestChange-package}
\alias{forestChange-package}
\alias{forestChange}
\docType{package}
\title{Computing Essential Biodiversity Variables from Global Forest Change Data}

\description{Metrics of two Essential Biodiversity Variables: forest extents and forest fragmentation indices (O'connor et al., 2015) <doi:10.1002/rse2.4> are computed by processing Global Forest Change data (Hansen et al., 2013) <doi:10.1126/science.1244693>. The Forest Change Data are extracted using either Geographic Administrative Units, see <https://gadm.org/>, or user-defined polygons. Most of the procedures can be understood implementing two functions: FCPolygon() and EBVmetric().}

\details{

The DESCRIPTION file:
\packageDESCRIPTION{forestChange}
\packageIndices{forestChange}
Maintainer: \packageMaintainer{forestChange}
}
\author{Wilson Lara <wilarhen@temple.edu>, Victor Gutierrez-Velez <victorhugo@temple.edu>}

\keyword{ package }


